// kbeatz-ui/src/api/generated/models/AlbumPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Album } from './Album';
export type AlbumPage = {
    content: Array<Album>;
    page: number;
    size: number;
    totalElements: number;
    totalPages: number;
};